pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apex-hackathon-adk"
version = "0.1.0"
description = "AI Agents built with Google ADK for hackathon"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.1.0",
    "python-dotenv>=0.19.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']


